Move out TTL map and webclient into their own packages, and restructure server packages
