+++ README.md
@@ -16,8 +16,8 @@ TODO
 * Server
     - [x] http server
       - [ ] Make a websocket server/client as an alternative to polling
-      - [ ] Move ttlmap out to a separate package that can be shared by http and websocket
-      - [ ] Move web client out of server directory to client directory as it will use both HTTP and websocket server depending on compile flags (or whatever)
+      - [x] Move ttlmap out to a separate package that can be shared by http and websocket
+      - [x] Move web client out of server directory to client directory as it will use both HTTP and websocket server depending on compile flags (or whatever)
         - [ ] Restructure the webservers so that we have a front line webserver for auth and static files, and for other requests it proxies to a distinct websocket or http webserver
       - [x] Instead of hanging indefinitely on GET sync/async, return with no update after timeout
       - [x] Instead of hanging indefinitely on GET match, return after a server timeout with http 204
