+++ internal/server/api/cache/ttlmap.go
@@ -1,4 +1,4 @@
-package webserver
+package cache
 
 // Credit to https://stackoverflow.com/questions/25484122/map-with-ttl-option-in-go
