+++ internal/server/api/http/static/index.html
