+++ internal/server/api/http/static/wasm_exec.js
