+++ internal/server/api/http/webserver_test.go
@@ -1,4 +1,4 @@
-package webserver
+package httpserver
 
 import (
 	"bytes"
