+++ internal/server/api/websocket/websocketserver.go
@@ -2,6 +2,7 @@ package websocketserver
 
 import (
 	"encoding/json"
+	"github.com/Ekotlikoff/gochess/internal/server/api/cache"
 	"github.com/Ekotlikoff/gochess/internal/server/match"
 	"github.com/gofrs/uuid"
 	"github.com/gorilla/websocket"
@@ -14,12 +15,12 @@ import (
 )
 
 var (
-	cache    *TTLMap
-	upgrader = websocket.Upgrader{} // use default options
+	sessionCache *cache.TTLMap
+	upgrader     = websocket.Upgrader{} // use default options
 )
 
 func init() {
-	cache = NewTTLMap(50, 1800, 10)
+	sessionCache = cache.NewTTLMap(50, 1800, 10)
 }
 
 type Credentials struct {
@@ -76,7 +77,7 @@ func StartSession(w http.ResponseWriter, r *http.Request) {
 	}
 	sessionTokenStr := sessionToken.String()
 	player := matchserver.NewPlayer(creds.Username)
-	cache.Put(sessionTokenStr, &player)
+	sessionCache.Put(sessionTokenStr, &player)
 	http.SetCookie(w, &http.Cookie{
 		Name:    "session_token",
 		Value:   sessionTokenStr,
@@ -98,7 +99,7 @@ func getSession(w http.ResponseWriter, r *http.Request) *matchserver.Player {
 		return nil
 	}
 	sessionToken := c.Value
-	player, err := cache.Get(sessionToken)
+	player, err := sessionCache.Get(sessionToken)
 	if err != nil {
 		log.Println("ERROR ", err)
 		w.WriteHeader(http.StatusInternalServerError)
